Refactored code to allow easier development
