--- firmware/check_icon.h
@@ -1,74 +1,0 @@
-// Image: 32 x 32
-uint16_t check_icon_palette[] = { 0x8000, 0x0400, 0x8410, 0x07E0, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
-
-const uint8_t check_icon_image[] = {
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x42, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x22, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x40, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x01, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x43, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x11, 0x33, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x10, 0x23, 0x31, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x10, 0x22, 0x33, 0x11, 0x11, 0x00, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x31, 0x11, 0x10, 0x22, 0x43, 0x31, 0x11, 0x10, 0x02, 0x22, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x33, 0x11, 0x10, 0x22, 0x44, 0x33, 0x11, 0x11, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x43, 0x11, 0x10, 0x44, 0x44, 0x43, 0x31, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x43, 0x33, 0x31, 0x44, 0x44, 0x44, 0x33, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x11, 0x00, 0x22, 0x24, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x11, 0x10, 0x02, 0x24, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x11, 0x02, 0x24, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x11, 0x02, 0x24, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x02, 0x24, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x14, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44
-};
-
-const uint8_t check_icon_mask[] = {
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0x8f, 0xff, 0xff,
-0xff, 0x07, 0xff, 0xff,
-0xfe, 0x03, 0xff, 0xff,
-0xfc, 0x01, 0xff, 0xff,
-0xf8, 0x00, 0xff, 0xff,
-0xf0, 0x00, 0x7f, 0xff,
-0xf0, 0x00, 0x3f, 0xff,
-0xf0, 0x00, 0x1f, 0xff,
-0xf0, 0x00, 0x0f, 0xff,
-0xf0, 0x00, 0x07, 0xff,
-0xf0, 0x00, 0x03, 0xff,
-0xf0, 0x00, 0x01, 0xff,
-0xf0, 0x00, 0x01, 0xff,
-0xf0, 0x08, 0x00, 0xff,
-0xf0, 0x0c, 0x00, 0x7f,
-0xf8, 0x3e, 0x00, 0x3f,
-0xf8, 0x3f, 0x00, 0x1f,
-0xff, 0xff, 0x80, 0x0f,
-0xff, 0xff, 0xc0, 0x07,
-0xff, 0xff, 0xe0, 0x07,
-0xff, 0xff, 0xf0, 0x07,
-0xff, 0xff, 0xf8, 0x07,
-0xff, 0xff, 0xfc, 0x07,
-0xff, 0xff, 0xfe, 0x07,
-0xff, 0xff, 0xff, 0x1f,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff
-};
-
-const ico_def_t check_icon = { 32, 32, check_icon_palette, check_icon_image, check_icon_mask };
