--- firmware/chip_icon.h
@@ -1,74 +1,0 @@
-// Image: 32 x 32
-uint16_t chip_icon_palette[] = { 0x0000, 0x8410, 0xC618, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
-
-const uint8_t chip_icon_image[] = {
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x02, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x22, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x22, 0x10, 0x20, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x23, 0x11, 0x01, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x21, 0x32, 0x21, 0x11, 0x10, 0x20, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x11, 0x21, 0x22, 0x22, 0x21, 0x11, 0x01, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00,
-0x00, 0x02, 0x11, 0x21, 0x22, 0x22, 0x22, 0x21, 0x11, 0x10, 0x20, 0x10, 0x10, 0x00, 0x10, 0x00,
-0x01, 0x22, 0x11, 0x22, 0x12, 0x22, 0x22, 0x22, 0x21, 0x11, 0x01, 0x10, 0x10, 0x00, 0x10, 0x00,
-0x01, 0x22, 0x31, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x10, 0x20, 0x10, 0x10, 0x00,
-0x01, 0x23, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x01, 0x10, 0x10, 0x00,
-0x01, 0x32, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x10, 0x20, 0x00,
-0x01, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x01, 0x00,
-0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x00,
-0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x00,
-0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x20, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x20, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x20, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x20, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-
-const uint8_t chip_icon_mask[] = {
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xf7, 0xff, 0xff,
-0xff, 0xe3, 0xff, 0xff,
-0xff, 0x83, 0x7f, 0xff,
-0xff, 0x02, 0x3f, 0xff,
-0xfe, 0x00, 0x37, 0xff,
-0xfe, 0x00, 0x23, 0xff,
-0xfc, 0x00, 0x03, 0x7f,
-0xf0, 0x00, 0x02, 0x3f,
-0xe0, 0x00, 0x00, 0x37,
-0xc0, 0x00, 0x00, 0x23,
-0x80, 0x00, 0x00, 0x03,
-0x80, 0x00, 0x00, 0x03,
-0x80, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x01,
-0xc0, 0x00, 0x00, 0x01,
-0xe0, 0x00, 0x00, 0x01,
-0xfc, 0x00, 0x00, 0x03,
-0xfe, 0x00, 0x00, 0x07,
-0xff, 0xc0, 0x00, 0x0f,
-0xff, 0xe0, 0x00, 0x1f,
-0xff, 0xfc, 0x00, 0x3f,
-0xff, 0xfe, 0x00, 0x7f,
-0xff, 0xff, 0xc0, 0xff,
-0xff, 0xff, 0xe1, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff
-};
-
-const ico_def_t chip_icon = { 32, 32, chip_icon_palette, chip_icon_image, chip_icon_mask };
