--- firmware/drum_icon1.h
@@ -1,74 +1,0 @@
-// Image: 32 x 32
-uint16_t drum_icon1_palette[] = { 0x0000, 0x0010, 0x001F, 0x0410, 0x8410, 0xC618, 0x07FF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
-
-const uint8_t drum_icon1_image[] = {
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x55, 0x55, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x05, 0x57, 0x77, 0x77, 0x77, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x57, 0x72, 0x22, 0x22, 0x21, 0x11, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x72, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x50, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x21, 0x15, 0x55, 0x55, 0x55, 0x44, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x15, 0x57, 0x77, 0x77, 0x77, 0x77, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x04, 0x57, 0x74, 0x44, 0x44, 0x44, 0x44, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x04, 0x70, 0x45, 0x55, 0x55, 0x55, 0x55, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x04, 0x40, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x04, 0x74, 0x45, 0x00, 0x05, 0x55, 0x55, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x40, 0x03, 0x66, 0x64, 0x44, 0x44, 0x77, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x36, 0x66, 0x33, 0x37, 0x77, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x03, 0x66, 0x63, 0x33, 0x44, 0x44, 0x44, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x36, 0x66, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x33, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-
-const uint8_t drum_icon1_mask[] = {
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xe0, 0x0f, 0xff,
-0xff, 0x80, 0x03, 0xff,
-0xff, 0x00, 0x01, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xff, 0x00, 0x01, 0xff,
-0xfc, 0x00, 0x03, 0xff,
-0xe0, 0x00, 0x0f, 0xff,
-0xc0, 0x7f, 0xff, 0xff,
-0xc3, 0xfe, 0x00, 0x0f,
-0xff, 0xfe, 0x00, 0x07,
-0xff, 0xfe, 0x00, 0x0f,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff
-};
-
-const ico_def_t drum_icon1 = { 32, 32, drum_icon1_palette, drum_icon1_image, drum_icon1_mask };
