--- firmware/error_icon.h
@@ -1,74 +1,0 @@
-// Image: 32 x 32
-uint16_t error_icon_palette[] = { 0x000F, 0x001F, 0x7BEF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
-
-const uint8_t error_icon_image[] = {
-0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x42, 0x22, 0x22, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x42, 0x20, 0x00, 0x00, 0x00, 0x02, 0x22, 0x22, 0x24, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x42, 0x00, 0x01, 0x11, 0x11, 0x11, 0x10, 0x00, 0x22, 0x22, 0x24, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x20, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x22, 0x22, 0x44, 0x44,
-0x44, 0x44, 0x40, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44,
-0x44, 0x44, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22, 0x24, 0x44,
-0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x22, 0x44,
-0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x22, 0x24,
-0x44, 0x01, 0x11, 0x11, 0x13, 0x11, 0x11, 0x11, 0x11, 0x11, 0x31, 0x11, 0x11, 0x10, 0x22, 0x24,
-0x40, 0x11, 0x11, 0x11, 0x33, 0x31, 0x11, 0x11, 0x11, 0x13, 0x33, 0x11, 0x11, 0x11, 0x02, 0x24,
-0x40, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x02, 0x22,
-0x40, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x02, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x10, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
-0x01, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x10, 0x24,
-0x01, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x10, 0x24,
-0x40, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x02, 0x24,
-0x40, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x02, 0x44,
-0x40, 0x11, 0x11, 0x11, 0x33, 0x31, 0x11, 0x11, 0x11, 0x13, 0x33, 0x11, 0x11, 0x11, 0x04, 0x44,
-0x44, 0x01, 0x11, 0x11, 0x13, 0x11, 0x11, 0x11, 0x11, 0x11, 0x31, 0x11, 0x11, 0x10, 0x24, 0x44,
-0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x44, 0x44,
-0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x04, 0x44, 0x44,
-0x44, 0x44, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x40, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x04, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x04, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x00, 0x01, 0x11, 0x11, 0x11, 0x10, 0x00, 0x44, 0x44, 0x44, 0x44, 0x44,
-0x44, 0x44, 0x44, 0x44, 0x44, 0x40, 0x00, 0x00, 0x00, 0x04, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44
-};
-
-const uint8_t error_icon_mask[] = {
-0xff, 0xf8, 0x07, 0xff,
-0xff, 0xc0, 0x00, 0xff,
-0xff, 0x80, 0x00, 0x7f,
-0xfe, 0x00, 0x00, 0x1f,
-0xfc, 0x00, 0x00, 0x0f,
-0xf8, 0x00, 0x00, 0x07,
-0xf0, 0x00, 0x00, 0x07,
-0xe0, 0x00, 0x00, 0x03,
-0xe0, 0x00, 0x00, 0x01,
-0xc0, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x00,
-0x80, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01,
-0x00, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x03,
-0x80, 0x00, 0x00, 0x07,
-0xc0, 0x00, 0x00, 0x07,
-0xe0, 0x00, 0x00, 0x0f,
-0xe0, 0x00, 0x00, 0x1f,
-0xf0, 0x00, 0x00, 0x3f,
-0xf8, 0x00, 0x00, 0x7f,
-0xfe, 0x00, 0x01, 0xff,
-0xff, 0x00, 0x03, 0xff,
-0xff, 0xe0, 0x1f, 0xff
-};
-
-const ico_def_t error_icon = { 32, 32, error_icon_palette, error_icon_image, error_icon_mask };
