+++ firmware/hardware.h
@@ -1,0 +1,10 @@
+#ifndef HARDWARE_H
+#define HARDWARE_H
+
+#include "pico/stdlib.h"
+
+void power_on();
+void power_off();
+void init_buttons_encoder();
+
+#endif //HARDWARE_H
