+++ firmware/icons.c
@@ -1,0 +1,655 @@
+// This file contains the definitions for all icon data.
+#include "gui.h"
+#include <stdint.h>
+
+// --- check_icon ---
+uint16_t check_icon_palette[] = { 0x8000, 0x0400, 0x8410, 0x07E0, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t check_icon_image[] = {
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x42, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x22, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x40, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x01, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x43, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x11, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x11, 0x33, 0x11, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x10, 0x23, 0x31, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x10, 0x22, 0x33, 0x11, 0x11, 0x00, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x31, 0x11, 0x10, 0x22, 0x43, 0x31, 0x11, 0x10, 0x02, 0x22, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x33, 0x11, 0x10, 0x22, 0x44, 0x33, 0x11, 0x11, 0x02, 0x22, 0x24, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x43, 0x11, 0x10, 0x44, 0x44, 0x43, 0x31, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x43, 0x33, 0x31, 0x44, 0x44, 0x44, 0x33, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x11, 0x00, 0x22, 0x22, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x11, 0x00, 0x22, 0x24, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x11, 0x10, 0x02, 0x24, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x11, 0x02, 0x24, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x11, 0x02, 0x24, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x31, 0x02, 0x24, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x33, 0x14, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44
+};
+const uint8_t check_icon_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0x8f, 0xff, 0xff,
+0xff, 0x07, 0xff, 0xff,
+0xfe, 0x03, 0xff, 0xff,
+0xfc, 0x01, 0xff, 0xff,
+0xf8, 0x00, 0xff, 0xff,
+0xf0, 0x00, 0x7f, 0xff,
+0xf0, 0x00, 0x3f, 0xff,
+0xf0, 0x00, 0x1f, 0xff,
+0xf0, 0x00, 0x0f, 0xff,
+0xf0, 0x00, 0x07, 0xff,
+0xf0, 0x00, 0x03, 0xff,
+0xf0, 0x00, 0x01, 0xff,
+0xf0, 0x00, 0x01, 0xff,
+0xf0, 0x08, 0x00, 0xff,
+0xf0, 0x0c, 0x00, 0x7f,
+0xf8, 0x3e, 0x00, 0x3f,
+0xf8, 0x3f, 0x00, 0x1f,
+0xff, 0xff, 0x80, 0x0f,
+0xff, 0xff, 0xc0, 0x07,
+0xff, 0xff, 0xe0, 0x07,
+0xff, 0xff, 0xf0, 0x07,
+0xff, 0xff, 0xf8, 0x07,
+0xff, 0xff, 0xfc, 0x07,
+0xff, 0xff, 0xfe, 0x07,
+0xff, 0xff, 0xff, 0x1f,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t check_icon = { 32, 32, check_icon_palette, check_icon_image, check_icon_mask };
+
+// --- chip_icon ---
+uint16_t chip_icon_palette[] = { 0x0000, 0x8410, 0xC618, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t chip_icon_image[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x02, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x22, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x10, 0x20, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x23, 0x11, 0x01, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x21, 0x32, 0x21, 0x11, 0x10, 0x20, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x11, 0x21, 0x22, 0x22, 0x21, 0x11, 0x01, 0x10, 0x10, 0x00, 0x10, 0x00, 0x00, 0x00,
+0x00, 0x02, 0x11, 0x21, 0x22, 0x22, 0x22, 0x21, 0x11, 0x10, 0x20, 0x10, 0x10, 0x00, 0x10, 0x00,
+0x01, 0x22, 0x11, 0x22, 0x12, 0x22, 0x22, 0x22, 0x21, 0x11, 0x01, 0x10, 0x10, 0x00, 0x10, 0x00,
+0x01, 0x22, 0x31, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x10, 0x20, 0x10, 0x10, 0x00,
+0x01, 0x23, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x01, 0x10, 0x10, 0x00,
+0x01, 0x32, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x10, 0x20, 0x00,
+0x01, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x01, 0x00,
+0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x00,
+0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x21, 0x00,
+0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x20, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x20, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x22, 0x20, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x22, 0x22, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x22, 0x20, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t chip_icon_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xf7, 0xff, 0xff,
+0xff, 0xe3, 0xff, 0xff,
+0xff, 0x83, 0x7f, 0xff,
+0xff, 0x02, 0x3f, 0xff,
+0xfe, 0x00, 0x37, 0xff,
+0xfe, 0x00, 0x23, 0xff,
+0xfc, 0x00, 0x03, 0x7f,
+0xf0, 0x00, 0x02, 0x3f,
+0xe0, 0x00, 0x00, 0x37,
+0xc0, 0x00, 0x00, 0x23,
+0x80, 0x00, 0x00, 0x03,
+0x80, 0x00, 0x00, 0x03,
+0x80, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x01,
+0xc0, 0x00, 0x00, 0x01,
+0xe0, 0x00, 0x00, 0x01,
+0xfc, 0x00, 0x00, 0x03,
+0xfe, 0x00, 0x00, 0x07,
+0xff, 0xc0, 0x00, 0x0f,
+0xff, 0xe0, 0x00, 0x1f,
+0xff, 0xfc, 0x00, 0x3f,
+0xff, 0xfe, 0x00, 0x7f,
+0xff, 0xff, 0xc0, 0xff,
+0xff, 0xff, 0xe1, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t chip_icon = { 32, 32, chip_icon_palette, chip_icon_image, chip_icon_mask };
+
+// --- drum_icon0 ---
+uint16_t drum_icon0_palette[] = { 0x0000, 0x0010, 0x001F, 0x0410, 0x8410, 0xC618, 0x07FF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t drum_icon0_image[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x55, 0x55, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x05, 0x57, 0x77, 0x77, 0x77, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x57, 0x72, 0x22, 0x22, 0x21, 0x11, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x72, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x50, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x21, 0x15, 0x55, 0x55, 0x55, 0x44, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x15, 0x57, 0x77, 0x77, 0x77, 0x77, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x57, 0x74, 0x44, 0x44, 0x44, 0x44, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x70, 0x45, 0x55, 0x55, 0x55, 0x55, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x40, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x74, 0x45, 0x55, 0x53, 0x00, 0x05, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x47, 0x74, 0x44, 0x43, 0x66, 0x60, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x04, 0x47, 0x77, 0x77, 0x33, 0x36, 0x66, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x44, 0x44, 0x44, 0x43, 0x33, 0x66, 0x60, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x36, 0x66, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x33, 0x30, 0x00,
+0x00, 0x36, 0x66, 0x66, 0x66, 0x66, 0x66, 0x60, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x03, 0x33, 0x33, 0x33, 0x33, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t drum_icon0_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xe0, 0x0f, 0xff,
+0xff, 0x80, 0x03, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xff, 0x80, 0x00, 0x7f,
+0xff, 0xe0, 0x00, 0x0f,
+0xff, 0xff, 0xfc, 0x07,
+0xe0, 0x01, 0xff, 0x87,
+0xc0, 0x00, 0xff, 0xff,
+0xe0, 0x01, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t drum_icon0 = { 32, 32, drum_icon0_palette, drum_icon0_image, drum_icon0_mask };
+
+// --- drum_icon1 ---
+uint16_t drum_icon1_palette[] = { 0x0000, 0x0010, 0x001F, 0x0410, 0x8410, 0xC618, 0x07FF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t drum_icon1_image[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x55, 0x55, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x05, 0x57, 0x77, 0x77, 0x77, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x57, 0x72, 0x22, 0x22, 0x21, 0x11, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x72, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x50, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x21, 0x15, 0x55, 0x55, 0x55, 0x44, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x15, 0x57, 0x77, 0x77, 0x77, 0x77, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x57, 0x74, 0x44, 0x44, 0x44, 0x44, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x70, 0x45, 0x55, 0x55, 0x55, 0x55, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x40, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x74, 0x45, 0x00, 0x05, 0x55, 0x55, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x40, 0x03, 0x66, 0x64, 0x44, 0x44, 0x77, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x36, 0x66, 0x33, 0x37, 0x77, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x03, 0x66, 0x63, 0x33, 0x44, 0x44, 0x44, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x36, 0x66, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x33, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t drum_icon1_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xe0, 0x0f, 0xff,
+0xff, 0x80, 0x03, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xfc, 0x00, 0x03, 0xff,
+0xe0, 0x00, 0x0f, 0xff,
+0xc0, 0x7f, 0xff, 0xff,
+0xc3, 0xfe, 0x00, 0x0f,
+0xff, 0xfe, 0x00, 0x07,
+0xff, 0xfe, 0x00, 0x0f,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t drum_icon1 = { 32, 32, drum_icon1_palette, drum_icon1_image, drum_icon1_mask };
+
+// --- drum_icon2 ---
+uint16_t drum_icon2_palette[] = { 0x0000, 0x0010, 0x001F, 0x0410, 0x8410, 0xC618, 0x07FF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t drum_icon2_image[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x55, 0x55, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x05, 0x57, 0x77, 0x77, 0x77, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x57, 0x72, 0x22, 0x22, 0x21, 0x11, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x72, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x50, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x21, 0x15, 0x55, 0x55, 0x55, 0x44, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x15, 0x57, 0x77, 0x77, 0x77, 0x77, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x57, 0x74, 0x44, 0x44, 0x44, 0x44, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x70, 0x45, 0x00, 0x35, 0x55, 0x55, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x40, 0x03, 0x66, 0x35, 0x55, 0x55, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x03, 0x66, 0x33, 0x53, 0x00, 0x05, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x03, 0x66, 0x33, 0x44, 0x43, 0x66, 0x63, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x03, 0x66, 0x33, 0x77, 0x77, 0x77, 0x33, 0x36, 0x66, 0x30, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x03, 0x66, 0x33, 0x00, 0x44, 0x44, 0x44, 0x44, 0x43, 0x33, 0x66, 0x63, 0x00, 0x00, 0x00,
+0x00, 0x03, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x36, 0x66, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x33, 0x30, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t drum_icon2_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xe0, 0x0f, 0xff,
+0xff, 0x80, 0x03, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfc, 0x00, 0x01, 0xff,
+0xf0, 0x00, 0x00, 0x7f,
+0xe0, 0xc0, 0x00, 0x0f,
+0xe3, 0xff, 0xfc, 0x07,
+0xff, 0xff, 0xff, 0x87,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t drum_icon2 = { 32, 32, drum_icon2_palette, drum_icon2_image, drum_icon2_mask };
+
+// --- drum_icon3 ---
+uint16_t drum_icon3_palette[] = { 0x0000, 0x0010, 0x001F, 0x0410, 0x8410, 0xC618, 0x07FF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t drum_icon3_image[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x55, 0x55, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x05, 0x57, 0x77, 0x77, 0x77, 0x55, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x57, 0x72, 0x22, 0x22, 0x21, 0x11, 0x55, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x72, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x50, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x22, 0x22, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x22, 0x21, 0x11, 0x11, 0x10, 0x00, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x21, 0x15, 0x55, 0x55, 0x55, 0x44, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x15, 0x57, 0x77, 0x77, 0x77, 0x77, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x57, 0x74, 0x44, 0x44, 0x44, 0x44, 0x77, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x74, 0x45, 0x55, 0x55, 0x30, 0x05, 0x44, 0x70, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x40, 0x55, 0x55, 0x55, 0x36, 0x63, 0x04, 0x40, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x04, 0x70, 0x45, 0x00, 0x05, 0x53, 0x36, 0x63, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x40, 0x03, 0x66, 0x64, 0x44, 0x43, 0x36, 0x63, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x36, 0x66, 0x33, 0x37, 0x77, 0x77, 0x43, 0x36, 0x63, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x03, 0x66, 0x63, 0x33, 0x44, 0x44, 0x44, 0x44, 0x00, 0x03, 0x36, 0x60, 0x00, 0x00,
+0x00, 0x03, 0x66, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x33, 0x00, 0x00,
+0x00, 0x33, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t drum_icon3_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xe0, 0x0f, 0xff,
+0xff, 0x80, 0x03, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xff, 0x00, 0x00, 0x7f,
+0xfc, 0x00, 0x00, 0x1f,
+0xe0, 0x00, 0x0e, 0x0f,
+0xc0, 0x7f, 0xff, 0x8f,
+0xc3, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t drum_icon3 = { 32, 32, drum_icon3_palette, drum_icon3_image, drum_icon3_mask };
+
+// --- error_icon ---
+uint16_t error_icon_palette[] = { 0x000F, 0x001F, 0x7BEF, 0xFFFF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t error_icon_image[] = {
+0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x42, 0x22, 0x22, 0x22, 0x24, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x42, 0x20, 0x00, 0x00, 0x00, 0x02, 0x22, 0x22, 0x24, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x42, 0x00, 0x01, 0x11, 0x11, 0x11, 0x10, 0x00, 0x22, 0x22, 0x24, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x20, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x22, 0x22, 0x44, 0x44,
+0x44, 0x44, 0x40, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x02, 0x22, 0x24, 0x44,
+0x44, 0x44, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22, 0x24, 0x44,
+0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x22, 0x44,
+0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x22, 0x24,
+0x44, 0x01, 0x11, 0x11, 0x13, 0x11, 0x11, 0x11, 0x11, 0x11, 0x31, 0x11, 0x11, 0x10, 0x22, 0x24,
+0x40, 0x11, 0x11, 0x11, 0x33, 0x31, 0x11, 0x11, 0x11, 0x13, 0x33, 0x11, 0x11, 0x11, 0x02, 0x24,
+0x40, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x02, 0x22,
+0x40, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x02, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x10, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x11, 0x10, 0x22,
+0x01, 0x11, 0x11, 0x11, 0x11, 0x33, 0x33, 0x33, 0x33, 0x33, 0x11, 0x11, 0x11, 0x11, 0x10, 0x24,
+0x01, 0x11, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x11, 0x10, 0x24,
+0x40, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x02, 0x24,
+0x40, 0x11, 0x11, 0x13, 0x33, 0x33, 0x11, 0x11, 0x11, 0x33, 0x33, 0x31, 0x11, 0x11, 0x02, 0x44,
+0x40, 0x11, 0x11, 0x11, 0x33, 0x31, 0x11, 0x11, 0x11, 0x13, 0x33, 0x11, 0x11, 0x11, 0x04, 0x44,
+0x44, 0x01, 0x11, 0x11, 0x13, 0x11, 0x11, 0x11, 0x11, 0x11, 0x31, 0x11, 0x11, 0x10, 0x24, 0x44,
+0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x02, 0x44, 0x44,
+0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x04, 0x44, 0x44,
+0x44, 0x44, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x40, 0x01, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x04, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x40, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x04, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x00, 0x01, 0x11, 0x11, 0x11, 0x10, 0x00, 0x44, 0x44, 0x44, 0x44, 0x44,
+0x44, 0x44, 0x44, 0x44, 0x44, 0x40, 0x00, 0x00, 0x00, 0x04, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44
+};
+const uint8_t error_icon_mask[] = {
+0xff, 0xf8, 0x07, 0xff,
+0xff, 0xc0, 0x00, 0xff,
+0xff, 0x80, 0x00, 0x7f,
+0xfe, 0x00, 0x00, 0x1f,
+0xfc, 0x00, 0x00, 0x0f,
+0xf8, 0x00, 0x00, 0x07,
+0xf0, 0x00, 0x00, 0x07,
+0xe0, 0x00, 0x00, 0x03,
+0xe0, 0x00, 0x00, 0x01,
+0xc0, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x00,
+0x80, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01,
+0x00, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x03,
+0x80, 0x00, 0x00, 0x07,
+0xc0, 0x00, 0x00, 0x07,
+0xe0, 0x00, 0x00, 0x0f,
+0xe0, 0x00, 0x00, 0x1f,
+0xf0, 0x00, 0x00, 0x3f,
+0xf8, 0x00, 0x00, 0x7f,
+0xfe, 0x00, 0x01, 0xff,
+0xff, 0x00, 0x03, 0xff,
+0xff, 0xe0, 0x1f, 0xff
+};
+const ico_def_t error_icon = { 32, 32, error_icon_palette, error_icon_image, error_icon_mask };
+
+// --- warn_icon ---
+uint16_t warn_icon_palette[] = { 0x0000, 0x0410, 0x8410, 0xC618, 0x07FF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t warn_icon_image[] = {
+0x00, 0x00, 0x02, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x00,
+0x00, 0x00, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x20,
+0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x22, 0x22,
+0x01, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x22,
+0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x22,
+0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x22,
+0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x00, 0x34, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x20,
+0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x40, 0x00, 0x04, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x20,
+0x01, 0x44, 0x44, 0x44, 0x44, 0x44, 0x40, 0x00, 0x04, 0x44, 0x44, 0x44, 0x44, 0x40, 0x22, 0x00,
+0x01, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x00, 0x34, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x00,
+0x00, 0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x20, 0x00,
+0x00, 0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x04, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00,
+0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x43, 0x03, 0x44, 0x44, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00,
+0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x41, 0x01, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x00, 0x00,
+0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x40, 0x00, 0x44, 0x44, 0x44, 0x44, 0x02, 0x20, 0x00, 0x00,
+0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x30, 0x00, 0x34, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00,
+0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x10, 0x00, 0x14, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x44, 0x30, 0x22, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x44, 0x02, 0x20, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x30, 0x22, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x30, 0x00, 0x34, 0x44, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x30, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x43, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t warn_icon_mask[] = {
+0xf8, 0x00, 0x00, 0x03,
+0xf0, 0x00, 0x00, 0x01,
+0xc0, 0x00, 0x00, 0x00,
+0x80, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x01,
+0x00, 0x00, 0x00, 0x01,
+0x80, 0x00, 0x00, 0x03,
+0x80, 0x00, 0x00, 0x03,
+0xc0, 0x00, 0x00, 0x07,
+0xc0, 0x00, 0x00, 0x07,
+0xe0, 0x00, 0x00, 0x0f,
+0xe0, 0x00, 0x00, 0x0f,
+0xf0, 0x00, 0x00, 0x1f,
+0xf0, 0x00, 0x00, 0x1f,
+0xf8, 0x00, 0x00, 0x3f,
+0xf8, 0x00, 0x00, 0x3f,
+0xfc, 0x00, 0x00, 0x7f,
+0xfc, 0x00, 0x00, 0x7f,
+0xfe, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x00, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xff, 0x00, 0x01, 0xff,
+0xff, 0x80, 0x03, 0xff,
+0xff, 0x80, 0x03, 0xff,
+0xff, 0xc0, 0x07, 0xff,
+0xff, 0xc0, 0x07, 0xff,
+0xff, 0xe0, 0x0f, 0xff,
+0xff, 0xe0, 0x1f, 0xff,
+0xff, 0xf0, 0x7f, 0xff,
+0xff, 0xf8, 0xff, 0xff
+};
+const ico_def_t warn_icon = { 32, 32, warn_icon_palette, warn_icon_image, warn_icon_mask };
+
+// --- x_icon ---
+uint16_t x_icon_palette[] = { 0x0000, 0x0010, 0x001F, 0x8410, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
+const uint8_t x_icon_image[] = {
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x33, 0x33, 0x30, 0x00, 0x00, 0x00, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x30, 0x00, 0x00, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x21, 0x10, 0x03, 0x33, 0x00, 0x00, 0x20, 0x33, 0x33, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x21, 0x11, 0x03, 0x33, 0x30, 0x02, 0x10, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x21, 0x11, 0x00, 0x33, 0x33, 0x02, 0x11, 0x00, 0x33, 0x33, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x21, 0x11, 0x10, 0x03, 0x33, 0x21, 0x11, 0x10, 0x03, 0x30, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x22, 0x11, 0x11, 0x00, 0x32, 0x11, 0x11, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x10, 0x01, 0x11, 0x11, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x02, 0x21, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x22, 0x11, 0x11, 0x11, 0x11, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x02, 0x21, 0x11, 0x11, 0x11, 0x10, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x22, 0x11, 0x11, 0x11, 0x03, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x11, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x11, 0x00, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x11, 0x10, 0x03, 0x33, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x11, 0x11, 0x11, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x12, 0x21, 0x11, 0x00, 0x33, 0x33, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x03, 0x22, 0x11, 0x10, 0x03, 0x33, 0x30, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x10, 0x30, 0x02, 0x21, 0x11, 0x00, 0x33, 0x33, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x10, 0x30, 0x00, 0x22, 0x11, 0x10, 0x03, 0x33, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x03, 0x00, 0x00, 0x02, 0x21, 0x11, 0x00, 0x30, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x02, 0x22, 0x11, 0x00, 0x00, 0x00, 0x00, 0x22, 0x11, 0x10, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x02, 0x22, 0x00, 0x00, 0x00, 0x00, 0x02, 0x21, 0x10, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x22, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+const uint8_t x_icon_mask[] = {
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xcf, 0xff, 0xff,
+0xff, 0x07, 0xf3, 0xff,
+0xff, 0x07, 0xe1, 0xff,
+0xfc, 0x03, 0xc0, 0xff,
+0xfc, 0x01, 0x80, 0x7f,
+0xfc, 0x00, 0x80, 0x3f,
+0xfc, 0x00, 0x00, 0x7f,
+0xfc, 0x00, 0x00, 0xff,
+0xfe, 0x00, 0x01, 0xff,
+0xfe, 0x00, 0x03, 0xff,
+0xff, 0x00, 0x07, 0xff,
+0xff, 0x80, 0x07, 0xff,
+0xff, 0xc0, 0x03, 0xff,
+0xff, 0xe0, 0x01, 0xff,
+0xff, 0xe0, 0x01, 0xff,
+0xff, 0xe0, 0x00, 0xff,
+0xff, 0xc0, 0x00, 0x7f,
+0xff, 0xc0, 0x00, 0x3f,
+0xff, 0x80, 0x00, 0x1f,
+0xff, 0x81, 0x80, 0x0f,
+0xff, 0x01, 0xc0, 0x0f,
+0xff, 0x03, 0xe0, 0x1f,
+0xfe, 0x07, 0xf0, 0x3f,
+0xff, 0x8f, 0xf8, 0x7f,
+0xff, 0xff, 0xfc, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff,
+0xff, 0xff, 0xff, 0xff
+};
+const ico_def_t x_icon = { 32, 34, x_icon_palette, x_icon_image, x_icon_mask };
