+++ firmware/icons.h
@@ -1,0 +1,52 @@
+#ifndef ICONS_H
+#define ICONS_H
+
+#include <stdint.h>
+#include "gui.h"
+
+extern uint16_t drum_icon0_palette[];
+extern const uint8_t drum_icon0_image[];
+extern const uint8_t drum_icon0_mask[];
+extern const ico_def_t drum_icon0;
+
+extern uint16_t drum_icon1_palette[];
+extern const uint8_t drum_icon1_image[];
+extern const uint8_t drum_icon1_mask[];
+extern const ico_def_t drum_icon1;
+
+extern uint16_t drum_icon2_palette[];
+extern const uint8_t drum_icon2_image[];
+extern const uint8_t drum_icon2_mask[];
+extern const ico_def_t drum_icon2;
+
+extern uint16_t drum_icon3_palette[];
+extern const uint8_t drum_icon3_image[];
+extern const uint8_t drum_icon3_mask[];
+extern const ico_def_t drum_icon3;
+
+extern uint16_t check_icon_palette[];
+extern const uint8_t check_icon_image[];
+extern const uint8_t check_icon_mask[];
+extern const ico_def_t check_icon;
+
+extern uint16_t chip_icon_palette[];
+extern const uint8_t chip_icon_image[];
+extern const uint8_t chip_icon_mask[];
+extern const ico_def_t chip_icon;
+
+extern uint16_t warn_icon_palette[];
+extern const uint8_t warn_icon_image[];
+extern const uint8_t warn_icon_mask[];
+extern const ico_def_t warn_icon;
+
+extern uint16_t x_icon_palette[];
+extern const uint8_t x_icon_image[];
+extern const uint8_t x_icon_mask[];
+extern const ico_def_t x_icon;
+
+extern uint16_t error_icon_palette[];
+extern const uint8_t error_icon_image[];
+extern const uint8_t error_icon_mask[];
+extern const ico_def_t error_icon;
+
+#endif
