+++ firmware/pio_patcher.h
@@ -6,8 +6,4 @@ struct pio_program *get_current_pio_program();
 void pio_patch_delays(const uint8_t *delays, uint8_t length);
 
 
-
-
-
-
 #endif
