+++ firmware/st7789.h
@@ -1,5 +1,6 @@
 #ifndef ST7789_H
 #define ST7789_H
+#include <stdbool.h>
 
 
 // Font definition table
