--- firmware/warn_icon.h
@@ -1,74 +1,0 @@
-// Image: 32 x 32
-uint16_t warn_icon_palette[] = { 0x0000, 0x0410, 0x8410, 0xC618, 0x07FF, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
-
-const uint8_t warn_icon_image[] = {
-0x00, 0x00, 0x02, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x00,
-0x00, 0x00, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x20,
-0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x22, 0x22,
-0x01, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x22,
-0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x22,
-0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x22,
-0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x00, 0x34, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x20,
-0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x40, 0x00, 0x04, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x20,
-0x01, 0x44, 0x44, 0x44, 0x44, 0x44, 0x40, 0x00, 0x04, 0x44, 0x44, 0x44, 0x44, 0x40, 0x22, 0x00,
-0x01, 0x44, 0x44, 0x44, 0x44, 0x44, 0x43, 0x00, 0x34, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x00,
-0x00, 0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x02, 0x20, 0x00,
-0x00, 0x14, 0x44, 0x44, 0x44, 0x44, 0x44, 0x04, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00,
-0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x43, 0x03, 0x44, 0x44, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00,
-0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x41, 0x01, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x00, 0x00,
-0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x40, 0x00, 0x44, 0x44, 0x44, 0x44, 0x02, 0x20, 0x00, 0x00,
-0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x30, 0x00, 0x34, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00,
-0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x10, 0x00, 0x14, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x44, 0x30, 0x22, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x44, 0x02, 0x20, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x00, 0x00, 0x04, 0x44, 0x30, 0x22, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x30, 0x00, 0x34, 0x44, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x44, 0x44, 0x30, 0x22, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x44, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x44, 0x44, 0x43, 0x02, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x40, 0x22, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x44, 0x44, 0x30, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x43, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-
-const uint8_t warn_icon_mask[] = {
-0xf8, 0x00, 0x00, 0x03,
-0xf0, 0x00, 0x00, 0x01,
-0xc0, 0x00, 0x00, 0x00,
-0x80, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x01,
-0x00, 0x00, 0x00, 0x01,
-0x80, 0x00, 0x00, 0x03,
-0x80, 0x00, 0x00, 0x03,
-0xc0, 0x00, 0x00, 0x07,
-0xc0, 0x00, 0x00, 0x07,
-0xe0, 0x00, 0x00, 0x0f,
-0xe0, 0x00, 0x00, 0x0f,
-0xf0, 0x00, 0x00, 0x1f,
-0xf0, 0x00, 0x00, 0x1f,
-0xf8, 0x00, 0x00, 0x3f,
-0xf8, 0x00, 0x00, 0x3f,
-0xfc, 0x00, 0x00, 0x7f,
-0xfc, 0x00, 0x00, 0x7f,
-0xfe, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x00, 0xff,
-0xff, 0x00, 0x01, 0xff,
-0xff, 0x00, 0x01, 0xff,
-0xff, 0x80, 0x03, 0xff,
-0xff, 0x80, 0x03, 0xff,
-0xff, 0xc0, 0x07, 0xff,
-0xff, 0xc0, 0x07, 0xff,
-0xff, 0xe0, 0x0f, 0xff,
-0xff, 0xe0, 0x1f, 0xff,
-0xff, 0xf0, 0x7f, 0xff,
-0xff, 0xf8, 0xff, 0xff
-};
-
-const ico_def_t warn_icon = { 32, 32, warn_icon_palette, warn_icon_image, warn_icon_mask };
