--- firmware/x_icon.h
@@ -1,78 +1,0 @@
-// Image: 32 x 34
-uint16_t x_icon_palette[] = { 0x0000, 0x0010, 0x001F, 0x8410, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000 };
-
-const uint8_t x_icon_image[] = {
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x33, 0x33, 0x30, 0x00, 0x00, 0x00, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x30, 0x00, 0x00, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x21, 0x10, 0x03, 0x33, 0x00, 0x00, 0x20, 0x33, 0x33, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x21, 0x11, 0x03, 0x33, 0x30, 0x02, 0x10, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x21, 0x11, 0x00, 0x33, 0x33, 0x02, 0x11, 0x00, 0x33, 0x33, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x21, 0x11, 0x10, 0x03, 0x33, 0x21, 0x11, 0x10, 0x03, 0x30, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x22, 0x11, 0x11, 0x00, 0x32, 0x11, 0x11, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x10, 0x01, 0x11, 0x11, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x02, 0x21, 0x11, 0x11, 0x11, 0x11, 0x11, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x22, 0x11, 0x11, 0x11, 0x11, 0x11, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x02, 0x21, 0x11, 0x11, 0x11, 0x10, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x22, 0x11, 0x11, 0x11, 0x03, 0x33, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x11, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x11, 0x00, 0x33, 0x30, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x11, 0x10, 0x03, 0x33, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x11, 0x11, 0x11, 0x03, 0x33, 0x30, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x12, 0x21, 0x11, 0x00, 0x33, 0x33, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x11, 0x03, 0x22, 0x11, 0x10, 0x03, 0x33, 0x30, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x02, 0x11, 0x10, 0x30, 0x02, 0x21, 0x11, 0x00, 0x33, 0x33, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x10, 0x30, 0x00, 0x22, 0x11, 0x10, 0x03, 0x33, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x21, 0x11, 0x03, 0x00, 0x00, 0x02, 0x21, 0x11, 0x00, 0x30, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x02, 0x22, 0x11, 0x00, 0x00, 0x00, 0x00, 0x22, 0x11, 0x10, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x02, 0x22, 0x00, 0x00, 0x00, 0x00, 0x02, 0x21, 0x10, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x22, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-
-const uint8_t x_icon_mask[] = {
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xcf, 0xff, 0xff,
-0xff, 0x07, 0xf3, 0xff,
-0xff, 0x07, 0xe1, 0xff,
-0xfc, 0x03, 0xc0, 0xff,
-0xfc, 0x01, 0x80, 0x7f,
-0xfc, 0x00, 0x80, 0x3f,
-0xfc, 0x00, 0x00, 0x7f,
-0xfc, 0x00, 0x00, 0xff,
-0xfe, 0x00, 0x01, 0xff,
-0xfe, 0x00, 0x03, 0xff,
-0xff, 0x00, 0x07, 0xff,
-0xff, 0x80, 0x07, 0xff,
-0xff, 0xc0, 0x03, 0xff,
-0xff, 0xe0, 0x01, 0xff,
-0xff, 0xe0, 0x01, 0xff,
-0xff, 0xe0, 0x00, 0xff,
-0xff, 0xc0, 0x00, 0x7f,
-0xff, 0xc0, 0x00, 0x3f,
-0xff, 0x80, 0x00, 0x1f,
-0xff, 0x81, 0x80, 0x0f,
-0xff, 0x01, 0xc0, 0x0f,
-0xff, 0x03, 0xe0, 0x1f,
-0xfe, 0x07, 0xf0, 0x3f,
-0xff, 0x8f, 0xf8, 0x7f,
-0xff, 0xff, 0xfc, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff,
-0xff, 0xff, 0xff, 0xff
-};
-
-const ico_def_t x_icon = { 32, 34, x_icon_palette, x_icon_image, x_icon_mask };
